+++ docs/CHANGE_LOG.md
@@ -8,6 +8,9 @@
 #### **BREAKING CHANGES:**
 - Renamed the DOM attribute "data-observe-options" to "data-observe-watch-options".
 
+#### **General Improvements:**
+- Minor internal refactoring to make things a little clearer.
+
 #### **Fixes:**
 - Fixed a fault that prevented the binding attribute from being removed from the DOM elements.
 - Fixed a fault that prevented changes from being detected when a DOM element is removed from the DOM.