+++ docs/CHANGE_LOG.md
@@ -14,6 +14,7 @@
 #### **Fixes:**
 - Fixed a fault that prevented the binding attribute from being removed from the DOM elements.
 - Fixed a fault that prevented changes from being detected when a DOM element is removed from the DOM.
+- Fixed some properties being assigned to internal watches when not required.
 
 <br>